update: 家事設定画面から家事登録画面を表示するようにする
